dmp/requirements.txt
pyyaml==6.0.1
pymysql==1.1.1
kafka-python==2.0.2
